separate osd viewer from control/nav UI
