+++ src/content-handlers/iiif/modules/uv-openseadragonbackgroundpanel-module/css/mobile.less
@@ -1,0 +1,10 @@
+// @media only screen and (min-width: @sm-min-width) {
+//     .uv-iiif-extension-host.uv-openseadragon-extension {
+//         // .centerPanel {
+
+//         //     .title {
+//         //         padding-top: @margin-small-vertical !important;
+//         //     }
+//         // }
+//     }
+// }
