+++ src/content-handlers/iiif/modules/uv-shared-module/css/styles.less
@@ -172,6 +172,13 @@
         });
   }
 
+  .backgroundPanel {
+    position: absolute;
+    overflow: hidden;
+    width: 100%;
+    height: 100%;
+  }
+
   .leftPanel,
   .rightPanel {
     background: @panel-dark-bg;
